+++ spec/http.spec.js
@@ -7,18 +7,21 @@ var requestor = require( 'request' ).defaults( { jar: true } );
 var metrics = require( 'cluster-metrics' );
 var port = 88988;
 var config = {
+		parseAhead: true,
 		port: port
 	};
 
 var authProvider, passport, middleware, http;
 
-describe( 'with http module', function() {
+describe( 'HTTP Module', function() {
 	var middlewareHit = [],
+		capturedParams = [],
 		request,
 		statusCode = 200,
 		response,
 		cleanup = function() {
 			middlewareHit = [];
+			capturedParams = [];
 			request = undefined;
 			response = undefined;
 			statusCode = 200;
@@ -33,15 +36,20 @@ describe( 'with http module', function() {
 		authProvider.users = {};
 		passport.resetUserCheck();
 
-		http.middleware( '/', function( req, res, next ) {
+		http.middleware( '/', function one( req, res, next ) {
+			if( req.params ) {
+				capturedParams.push( req.param( 'one' ) );
+				capturedParams.push( req.preparams.two );
+			}
+			console.log( capturedParams );
 			middlewareHit.push( 1 );
 			next();
 		} );
-		http.middleware( '/', function( req, res, next ) {
+		http.middleware( '/', function two( req, res, next ) {
 			middlewareHit.push( 2 );
 			next();
 		} );
-		http.middleware( '/', function( req, res, next ) {
+		http.middleware( '/', function three( req, res, next ) {
 			middlewareHit.push( 3 );
 			next();
 		} );
@@ -126,6 +134,10 @@ describe( 'with http module', function() {
 			middlewareHit.should.eql( [ 1, 2, 3 ] );
 		} );
 
+		it( 'should have captured parameters in middleware', function() {
+			capturedParams.should.eql( [ 'a', 'b' ] );
+		} );
+
 		after( cleanup );
 	} );
 
@@ -169,6 +181,98 @@ describe( 'with http module', function() {
 			authProvider.authenticate = originalAuthenticate;
 			authProvider.users = {};
 		} );
+	} );
+
+	after( function() {
+		http.stop();
+	} );
+} );
+
+describe( 'HTTP Routing Precedence', function() {
+	var middlewareHit = [],
+		request,
+		statusCode = 200,
+		response,
+		cleanup = function() {
+			middlewareHit = [];
+			request = undefined;
+			response = undefined;
+			statusCode = 200;
+		};
+
+	before( function() {
+		var specialConfig = {
+			port: port,
+			static: './spec/public'
+		};
+		middleware = require( '../src/http/middleware.js' )( config, metrics );
+		http = require( '../src/http/http.js' )( specialConfig, requestor, undefined, middleware, metrics );
+		http.middleware( '/thing', function( req, res, next ) {
+			next();
+		} );
+		http.route( /^\/thing/, 'all', function( req, res ) {
+			request = req;
+			res.status( statusCode ).send( response );
+		} );
+		http.start();
+	} );
+
+	describe( 'when getting a static file', function() {
+		var result;
+
+		before( function( done ) {
+			requestor.get( {
+				url: 'http://localhost:88988/txt/hello.txt' 
+			}, function( err, resp ) {
+				result = resp.body;
+				done();
+			} );
+		} );
+
+		it( 'should return file contents', function() {
+			result.should.equal( 'hello, world!' );
+		} );
+
+		after( cleanup );
+	} );
+
+	describe( 'when requesting a missing static file', function() {
+		var result;
+
+		before( function( done ) {
+			requestor.get( {
+				url: 'http://localhost:88988/txt/missing.txt'
+			}, function( err, resp ) {
+				result = resp.body;
+				done();
+			} );
+		} );
+
+		it( 'should report files as unavailable', function() {
+			result.should.equal( 'Cannot GET /txt/missing.txt\n' );
+		} );
+
+		after( cleanup );
+	} );
+
+	describe( 'when calling get on route', function() {
+		var result;
+
+		before( function( done ) {
+			response = 'hey, a thing';
+			requestor.get( {
+				url: 'http://localhost:88988/thing/a/b?c=3' 
+			}, function( err, resp ) {
+				result = resp;
+				done();
+			} );
+		} );
+
+		it( 'should return file contents', function() {
+			result.body.should.equal( 'hey, a thing' );
+		} );
+
+		after( cleanup );
 	} );
 
 	after( function() {