+++ spec/regexPath.spec.js
@@ -79,3 +79,63 @@ describe( 'when pattern does not contain ^', function() {
 		url3Match.should.be.false; // jshint ignore:line
 	} );
 } );
+
+describe( 'when pattern contains leading forward slash after ^', function() {
+	var prefix = '/api';
+	var url1 = '/api/views/this/is/dumb';
+	var url2 = '/board/10/user';
+	var url3 = '/api/user';
+	var match = /^\/views\//;
+	var url1Match, url2Match, url3Match;
+
+	before( function() {
+		var pattern = regex.prefix( prefix, match );
+		console.log( pattern );
+		url1Match = pattern.test( url1 );
+		url2Match = pattern.test( url2 );
+		url3Match = pattern.test( url3 );
+	} );
+
+
+	it( 'it should match URLs beginning with the correct prefix', function() {
+		url1Match.should.be.true; // jshint ignore:line
+	} );
+
+	it( 'it should not match URLs missing the prefix', function() {
+		url2Match.should.be.false; // jshint ignore:line
+	} );
+
+	it( 'it should not match URLs that are partial matches', function() {
+		url3Match.should.be.false; // jshint ignore:line
+	} );
+} );
+
+describe( 'when pattern contains leading forward slash without ^', function() {
+	var prefix = '/api';
+	var url1 = '/api/test/user/this/is/dumb';
+	var url2 = '/board/10/user';
+	var url3 = '/api/user';
+	var match = /\/user\//;
+	var url1Match, url2Match, url3Match;
+
+	before( function() {
+		var pattern = regex.prefix( prefix, match );
+		console.log( pattern );
+		url1Match = pattern.test( url1 );
+		url2Match = pattern.test( url2 );
+		url3Match = pattern.test( url3 );
+	} );
+
+
+	it( 'it should match URLs beginning with the correct prefix', function() {
+		url1Match.should.be.true; // jshint ignore:line
+	} );
+
+	it( 'it should not match URLs missing the prefix', function() {
+		url2Match.should.be.false; // jshint ignore:line
+	} );
+
+	it( 'it should not match URLs that are partial matches', function() {
+		url3Match.should.be.false; // jshint ignore:line
+	} );
+} );